Breaking apart mdspan/mdarray into host_ and device_ variants
